requirements.txt
Kivy==2.3.1
requests==2.32.5
python-dotenv==1.0.0
